// packages/bespoke/game/Graphical/VickreyAuction3/src/interface.d.ts
export interface ICreateParams {

    /** CreateParams round */
    round?: (number|null);

    /** CreateParams groupSize */
    groupSize?: (number|null);

    /** CreateParams startingPrice */
    startingPrice?: (number|null);

    /** CreateParams privatePrice */
    privatePrice?: (string[]|null);

    /** CreateParams nextPhaseKey */
    nextPhaseKey?: (string|null);
}